requirements.txt
PyYAML
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailure
allure-pytest